Optimized architecture layout for bytecode use.
Moved the run loop from the RobotPlayer to the Bot's themselves.  This saves a couple bytecodes as there isn't an unnesscary function call.  Also makes the code prettier :)
Added some functionality to Debugger class.  We can now find out the code cost of sections of code using the counter functions.