BC hacks shave off a few more BC.
Lost another 2 BC per bot in each of the inner loops. Down to ~105 per
bot overall. (I have a feeling that the difference might be slightly
more when there are ground units around since the pair of conditionals
in the second loop didn't come into effect in the first run of profiles
on this section.)