Sensing and enemy prioritizing
These values may need to be optimized, but they should work for now. BC
Usage has bumped up a bit with friendly units requiring ~113/unit now
due to extra conditionals in the code. Enemies are up to ~126 which
includes the cost of the prioritizing function call.

This code will work best when the attacked units health is updated at
the end of each individual robot's turn. That way, when we do our scan
robots are correctly prioritized based on current health. Otherwise,
we'll need to do some sort of complicated comms shit to make sure the
surrounding bots know which target we have and what his expected energon
should be if we score a hit. And they'll have to do some uglyness with
the messages and take that into account when they're calculating
priorities... Ergh.